+++ app/src/main/java/org/mozilla/tv/firefox/navigationoverlay/NavigationOverlayFragment.kt
@@ -23,15 +23,12 @@ import android.view.LayoutInflater
 import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageButton
 import android.widget.ScrollView
 import android.widget.Toast
 import kotlinx.android.synthetic.main.browser_overlay.*
 import kotlinx.android.synthetic.main.browser_overlay_top_nav.*
 import kotlinx.android.synthetic.main.pocket_video_mega_tile.*
 import kotlinx.coroutines.Job
-import mozilla.components.browser.domains.autocomplete.ShippedDomainsProvider
-import mozilla.components.support.ktx.android.view.hideKeyboard
 import org.mozilla.tv.firefox.MainActivity
 import org.mozilla.tv.firefox.R
 import org.mozilla.tv.firefox.experiments.ExperimentConfig
@@ -50,14 +47,8 @@ import org.mozilla.tv.firefox.settings.SettingsFragment
 import org.mozilla.tv.firefox.telemetry.TelemetryIntegration
 import org.mozilla.tv.firefox.telemetry.UrlTextInputLocation
 import org.mozilla.tv.firefox.utils.ServiceLocator
-import org.mozilla.tv.firefox.utils.URLs
-import org.mozilla.tv.firefox.utils.ViewUtils
-import org.mozilla.tv.firefox.widget.IgnoreFocusMovementMethod
 import org.mozilla.tv.firefox.widget.InlineAutocompleteEditText
 import java.lang.ref.WeakReference
-
-private const val NAVIGATION_BUTTON_ENABLED_ALPHA = 1.0f
-private const val NAVIGATION_BUTTON_DISABLED_ALPHA = 0.3f
 
 private const val SHOW_UNPIN_TOAST_COUNTER_PREF = "show_upin_toast_counter"
 private const val MAX_UNPIN_TOAST_COUNT = 3
@@ -85,7 +76,7 @@ enum class NavigationEvent {
 }
 
 @Suppress("LargeClass") // TODO remove this. See https://github.com/mozilla-mobile/firefox-tv/issues/1187
-class NavigationOverlayFragment : Fragment(), View.OnClickListener {
+class NavigationOverlayFragment : Fragment() {
     companion object {
         const val FRAGMENT_TAG = "overlay"
     }
@@ -129,8 +120,6 @@ class NavigationOverlayFragment : Fragment(), View.OnClickListener {
     private var currFocus: View? = null
         get() = activity?.currentFocus
 
-    private var hasUserChangedURLSinceEditTextFocused = false
-
     private lateinit var serviceLocator: ServiceLocator
     private lateinit var toolbarViewModel: ToolbarViewModel
     private lateinit var pinnedTileViewModel: PinnedTileViewModel
@@ -157,13 +146,14 @@ class NavigationOverlayFragment : Fragment(), View.OnClickListener {
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        topNavContainer.forEachChild {
-            it.nextFocusDownId = navUrlInput.id
-            if (it.isFocusable) it.setOnClickListener(this)
-        }
+        ToolbarUiController(
+            toolbarViewModel,
+            ::exitFirefox,
+            { currFocus },
+            ::updateFocusableViews,
+            onNavigationEvent
+        ).onCreateView(view, viewLifecycleOwner, fragmentManager!!)
 
-        setupUrlInput()
-        initToolbar()
         initMegaTile()
         initPinnedTiles()
 
@@ -191,92 +181,6 @@ class NavigationOverlayFragment : Fragment(), View.OnClickListener {
         registerForContextMenu(tileContainer)
 
         updateFocusableViews()
-    }
-
-    private fun setupUrlInput() = with(navUrlInput) {
-        setOnCommitListener {
-            val userInput = text.toString()
-            if (userInput == URLs.APP_URL_HOME) {
-                // If the input points to home, we short circuit and hide the keyboard, returning
-                // the user to the home screen
-                this.hideKeyboard()
-                return@setOnCommitListener
-            }
-
-            if (userInput.isNotEmpty()) {
-                val cachedAutocompleteResult = lastAutocompleteResult // setText clears the reference so we cache it here.
-                setText(cachedAutocompleteResult.text)
-                onNavigationEvent.invoke(NavigationEvent.LOAD_URL, userInput, cachedAutocompleteResult)
-            }
-        }
-        this.movementMethod = IgnoreFocusMovementMethod()
-        val autocompleteProvider = ShippedDomainsProvider().apply {
-            initialize(
-                    context = context
-            )
-        }
-        setOnFilterListener { searchText, view ->
-            val result = autocompleteProvider.getAutocompleteSuggestion(searchText)
-            if (result != null)
-                view?.onAutocomplete(InlineAutocompleteEditText.AutocompleteResult(result.text, result.source, result.totalItems))
-        }
-
-        setOnUserInputListener { hasUserChangedURLSinceEditTextFocused = true }
-        setOnFocusChangeListener { _, hasFocus -> if (!hasFocus) hasUserChangedURLSinceEditTextFocused = false }
-    }
-
-    private fun initToolbar() {
-        fun updateOverlayButtonState(isEnabled: Boolean, overlayButton: ImageButton) {
-            overlayButton.isEnabled = isEnabled
-            overlayButton.isFocusable = isEnabled
-            overlayButton.alpha =
-                    if (isEnabled) NAVIGATION_BUTTON_ENABLED_ALPHA else NAVIGATION_BUTTON_DISABLED_ALPHA
-        }
-
-        toolbarViewModel.state.observe(viewLifecycleOwner, Observer {
-            if (it == null) return@Observer
-            val focusedView = currFocus
-            updateOverlayButtonState(it.backEnabled, navButtonBack)
-            updateOverlayButtonState(it.forwardEnabled, navButtonForward)
-            updateOverlayButtonState(it.pinEnabled, pinButton)
-            updateOverlayButtonState(it.refreshEnabled, navButtonReload)
-            updateOverlayButtonState(it.desktopModeEnabled, desktopModeButton)
-
-            updateFocusableViews(focusedView)
-
-            pinButton.isChecked = it.pinChecked
-            desktopModeButton.isChecked = it.desktopModeChecked
-            turboButton.isChecked = it.turboChecked
-
-            if (!hasUserChangedURLSinceEditTextFocused) {
-                // The url can get updated in the background, e.g. if a loading page is redirected. We
-                // don't want a url update to interrupt the user typing so we don't update the url from
-                // the background if the user has already updated the url themselves.
-                //
-                // We revert this state when the view is unfocused: it ensures the URL is usually accurate
-                // (for security reasons) and it's simple compared to other options which keep more state.
-                //
-                // One problem this solution has is that if the URL is updated in the background rapidly,
-                // sometimes key events will be dropped, but I don't think there's much we can do about this:
-                // we can't determine if the keyboard is up or not and focus isn't a good indicator because
-                // we can focus the EditText without opening the soft keyboard and the user won't even know
-                // these are inaccurate!
-                navUrlInput.setText(it.urlBarText)
-            }
-        })
-
-        toolbarViewModel.events.observe(viewLifecycleOwner, Observer {
-            it?.consume {
-                when (it) {
-                    is ToolbarViewModel.Action.ShowTopToast -> ViewUtils.showCenteredTopToast(context, it.textId)
-                    is ToolbarViewModel.Action.ShowBottomToast -> ViewUtils.showCenteredBottomToast(context, it.textId)
-                    is ToolbarViewModel.Action.SetOverlayVisible -> serviceLocator.screenController
-                            .showNavigationOverlay(fragmentManager, it.visible)
-                    ToolbarViewModel.Action.ExitFirefox -> exitFirefox()
-                }.forceExhaustive
-                true
-            }
-        })
     }
 
     private fun exitFirefox() {
@@ -311,7 +215,10 @@ class NavigationOverlayFragment : Fragment(), View.OnClickListener {
     }
 
     private fun initMegaTile() {
-        pocketVideoMegaTileView.setOnClickListener(this)
+        pocketVideoMegaTileView.setOnClickListener { view ->
+            val event = NavigationEvent.fromViewClick(view.id) ?: return@setOnClickListener
+            onNavigationEvent.invoke(event, null, null)
+        }
 
         pocketViewModel.state.observe(viewLifecycleOwner, Observer { state ->
             when (state) {
@@ -384,23 +291,6 @@ class NavigationOverlayFragment : Fragment(), View.OnClickListener {
             val marginLayoutParams = it as ViewGroup.MarginLayoutParams
             marginLayoutParams.bottomMargin = -tileBottomMargin
         }
-    }
-
-    override fun onClick(view: View?) {
-        val event = NavigationEvent.fromViewClick(view?.id)
-                ?: return
-
-        when (event) {
-            NavigationEvent.BACK -> toolbarViewModel.backButtonClicked()
-            NavigationEvent.FORWARD -> toolbarViewModel.forwardButtonClicked()
-            NavigationEvent.RELOAD -> toolbarViewModel.reloadButtonClicked()
-            NavigationEvent.PIN_ACTION -> toolbarViewModel.pinButtonClicked()
-            NavigationEvent.TURBO -> toolbarViewModel.turboButtonClicked()
-            NavigationEvent.DESKTOP_MODE -> toolbarViewModel.desktopModeButtonClicked()
-            NavigationEvent.EXIT_FIREFOX -> toolbarViewModel.exitFirefoxButtonClicked()
-            else -> Unit // Nothing to do.
-        }
-        onNavigationEvent.invoke(event, null, null)
     }
 
     override fun onCreateContextMenu(menu: ContextMenu?, v: View?, menuInfo: ContextMenu.ContextMenuInfo?) {
